Use a username with whispers and mod chats
“username to username:” or “username to mods:” was originally prefixed
to the chat and then the whole string was sent as a server message.
Now, “username to username:” is considered the user’s name for the
chat. This way, clients will process it as a user chat instead of a
server message (this results in the client coloring the chat
differently).