Fix builder bug that allowed ordering disabled units
If a unit was ordered while it was enabled, and then the order was
canceled after the unit was disabled, the order button would appear and
allow the disabled unit to be ordered more. This would continue to work
as long as there was at least one order of the disabled unit in the
queue to cancel.